+++ Vials.Core.UnitTests/Utils/ColorCounter.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Text;
-using Vials.Shared;
+using Vials.Shared.Objects;
 
 namespace Vials.Core.UnitTests.Utils
 {
