+++ Vials.Core/ISetGenerator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using Vials.Shared;
+using Vials.Shared.Objects;
 
 namespace Vials.Core
 {
