+++ Vials.Core/PathFinder.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Vials.Shared;
+using Vials.Shared.Objects;
 
 namespace Vials.Core
 {
