+++ Vials.Shared.UnitTests/VialSetHistoryTests.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Vials.Shared.Extensions;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared.UnitTests
 {
