+++ Vials.Shared.UnitTests/VialSetTests.cs
@@ -3,6 +3,7 @@ using Shouldly;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared.UnitTests
 {
