+++ Vials/Client/CodeBehind/VialSetView.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Vials.Client.Shared;
 using Vials.Shared;
+using Vials.Shared.Objects;
 using Vials.Shared.Components;
 using Vials.Shared.Events;
