+++ Vials/Client/Service/PathService.cs
@@ -6,10 +6,10 @@ using System.Net.Http.Json;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
+using Vials.Shared.Objects;
 
-namespace Vials.Shared.Service
+namespace Vials.Client.Service
 {
-    // TODO move to client
     public class PathService : IPathService
     {
         private readonly HttpClient _http;
