+++ Vials/Server/Controllers/PathController.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Vials.Core;
-using Vials.Shared;
+using Vials.Shared.Objects;
 
 namespace Vials.Server.Controllers
 {
