+++ Vials/Shared/Cloner.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared
 {
