+++ Vials/Shared/Components/IDefault.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared.Components
 {
