+++ Vials/Shared/Events/VialClickedHandler.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Vials.Shared;
 using Vials.Shared.Components;
 using Vials.Shared.Events;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared.Events
 {
