+++ Vials/Shared/ICloner.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared
 {
