+++ Vials/Shared/Objects/Color.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Vials.Shared
+namespace Vials.Shared.Objects
 {
     public enum Color
     {
