+++ Vials/Shared/Objects/Pouring.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Vials.Shared
+namespace Vials.Shared.Objects
 {
     public class Pouring
     {
