+++ Vials/Shared/Objects/VialSet.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Vials.Shared
+namespace Vials.Shared.Objects
 {
     public class VialSet
     {
