+++ Vials/Shared/VialSetHandler.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Vials.Shared.Extensions;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared
 {
