+++ Vials/Shared/VialSetHistory.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Vials.Shared.Objects;
 
 namespace Vials.Shared
 {
